switch to double in the first place

+++ aslam_cv/aslam_cameras/include/aslam/cameras/implementation/PinholeProjection.hpp
@@ -780,13 +780,13 @@ bool PinholeProjection<DISTORTION_T>::initializeIntrinsics(const std::vector<Gri
 
   // Get the median of the guesses if available.
   if(f_guesses.empty()) {
-   float input_guess;
+   double input_guess;
    std::cout << "Initialization of focal length failed. Provide manual initialization: ";
    std::cin >> input_guess;
-   SM_ASSERT_GT(std::runtime_error, input_guess, 0.0f, 
+   SM_ASSERT_GT(std::runtime_error, input_guess, 0.0, 
                 "Focal length needs to be positive.");
    std::cout << "Initializing focal length to " << input_guess << "\n";
-   f_guesses.push_back(static_cast<double>(input_guess));
+   f_guesses.push_back(input_guess);
   }
   double f0 = PinholeHelpers::medianOfVectorElements(f_guesses);
